Panzer: Fixing SG tests to reflect changes in ETI
I was compiling the somewhat useless "short" LID version of the DOFManager
and linear object factories. This was basically to demonstrate that these
things could function using something other then int. It worked fine but
became painful to handle and maintain. This should be the last of the required
fixes.

Also had to add buildLocalIds to the call for the UGI in the ugi rig.
Fixed a stupid bug as well.