Teuchos: ParameterList: Cleaned up the documentation a bit.
Clang likes to tell you about broken HTML in Doxygen documentation.
I fixed some of this in ParameterList's documentation, and took the
opportunity to rewrite the public documentation for the class and
some commonly used methods.  It's important that ParameterList, a
commonly used class by most of Trilinos and its users, have clear
documentation that clearly defines the class' semantics.