Tpetra: Broke up CrsMatrix unit tests into separate files.
CrsMatrix_UnitTests.cpp was taking a really long time to compile.
There seems to be some issue with GCC that makes compilation time
blow up beyond a certain threshold of template instantiation.  We
haven't yet figured out how to work around the GCC issue.  Another
way to deal with it, though, is to expose parallelism in the build:
that is, break up the long file into separate files, each with one
or a few closely related unit tests.  I've begun doing this for
CrsMatrix_UnitTests.cpp, and it helps a lot with compilation time.
It used to take several minutes for GCC 4.7.2 to build that file on
a Mac with 4 cores and 16 GB of RAM; now it takes less than a minute.

Splitting up the long file is also a good idea for making it easier
to find and edit a particular unit test.  If someone wants to split
CrsMatrix_UnitTests.cpp further, they should choose file names that
suggest the topic of the tests in the file.  The files I've added in
this commit give examples, such as "LeftRightScale" for left and right
diagonal scaling, and "TriSolve" for triangular solve.