Fix a stk-mesh hotspot (Entity::operator==).
nrg team profiling of an adagio contact problem (p003_cancrush_hex) revealed
a lot of time being spent in the comparison operator for Entity during the
contact checkFaceSharing function. The Entity comparison operator was doing
checks on 'is_valid' and then accessing and comparing the entity-key. All it really
needs to do is compare the pointer-to-EntityImpl. Entities are equal iff those
pointers are the same (is_valid doesn't need to be checked). Fixing this removes ~40s from the
checkFaceSharing run time, and lowers the run-time for the cancrush problem from
 ~14:04 to ~13:12 (mm:ss).

(From Sierra)