Updated build system to work even in presence of python 3
PyTrilinos is not yet upgraded to work under python 3, but if a user's
environment is set so that "python" executes python 3, then the
scripts in the util directory won't work.  This commit changes those
scripts to be copied to the build directory, and rewritten to ensure
that they are executed by ${PYTHON_EXECUTABLE} instead of the default
environment python.

To support this, the util directory had to be added as a subdirectory
in the main PyTrilinos CMakeLists.txt file.  The
cmake/PyTrilinosMakeTest.cmake file had to be updated to run the
scripts from the build directory, not the source directory (which are
scripts that no longer exist).  The CMakeLists.txt files in the
example and test directories had to be modified in the same way for
copying testing utility files.