Make framework depend on stk-io...
So that a framework unit-test can demonstrate reading an exodus file
into framework using frio, then pull out the stk-mesh object from framework
and write it to a different exodus file using stk-io. This is a
use-case that the nrg team attempted and hit some snags that
are fixed in this commit.

This commit also adds to stk-mesh the knowledge of which field is the
coordinate field so that the client code doesn't need to know the
particular string that was used to declare the coord field.
Now stk-io and IossBridge use that coord-field-getter rather than using
a string-name for coordinates.

(From Sierra)