Updated the backlog of CMake/CTest/CDash issues the remain.
I deleted a lot of items that I know are fixed (which is great) and there are
some more that might be fixed but I would have to check.

I deleted the "Done" section since I don't see future value in keeping this.
Git knows the prior versions of this file in case someone wants to see it.