Kokkos,Tpetra: MultiVector: elementWiseMultiply now overwrites NaNs.
Deaglan Halligan (Purdue) found that Tpetra::MultiVector's
elementWiseMultiply() method was allowing NaN values to persist in the
left-hand side of the expression, even when its coefficient is zero.
This does not match the semantics of other MultiVector methods, such
as update() (both variants) and multiply().

This commit fixes Tpetra::MultiVector::elementWiseMultiply to have
overwrite semantics, so that if the coefficient of *this is zero,
the method overwrites the entries of *this, even if some of them are
NaN.  It also documents the overwrite semantics of Kokkos' GEMM() and
Tpetra's multiply().

Build/Test Cases Summary
Enabled Packages: Teuchos, Kokkos, Tpetra, Ifpack2, Belos, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=286,notpassed=0 (12.29 min)
3) SERIAL_RELEASE_SS => passed: passed=267,notpassed=0 (11.72 min)
4) MPI_RELEASE_SS => passed: passed=286,notpassed=0 (15.45 min)