Piro: Enabled customizable default provider
An ExtensibleFactory used to return a null pointer when no suitable
provider was found. This behavior is now customizable at run-time.
The implementation reuses the Provider concepts and machinery.

In addition, the new feature is exploited in SolverFactory to simplify
the common case where only one type of auxiliary object is required.