Implementing a more efficient matrix transpose.
This is based on the approach in EpetraExt (written by Alan Williams?).
I've made a few optimizations (like using exact nonzero layout and
static profiling where possible), but the basic approach is the same.

There are probably better approaches, but the previous implementation
was putting the entire job of transposing on fillComplete.