Teuchos: TypeNameTraits: Updated exception msg on failed demangling.
When name demangling is enabled, Teuchos tests in a debug build
whether demangling succeeded, and raises an exception if it did not.
The exception message tells users how to configure Trilinos to disable
name demangling, so that the exception won't be thrown anymore.

The exception message referred to a now defunct GNU Autotools option,
rather than to the CMake option.  I corrected the message so that it
refers to the CMake option, and patiently explains to users how to set
the option.