Fixing print message when pull fails
The script was reporting that files were pulled when a pull failed (because
'Already up-to-date' was missing).  This was confusing but the script still
looked at the return value when deciding that a pull failed so it still did
the right thing.

With this change, I have fixed the output to show the correct status of each
pull and added unit tests to demonstrate and production the expected output.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=102,notpassed=0 (1.79 min)
1) SERIAL_RELEASE => passed: passed=102,notpassed=0 (1.11 min)