Allowing ${PROJECT_NAME}_IGNORE_MISSING_EXTRA_REPOSITORIES to be overridden
I set this to allow for easier local testing of automated builds.  This way, I
don't need to clone all of the extra repos that are set in the
ExtraRepositoriesList.cmake file.  I did this to simplify debugging of issues
with a VERA build (that has like 22 extra repos now total of which I don't
have them all cloned).