Disabling SuperLU for the installation testing
The install for the installation testing isn't built with superlu so building
the testing against it with superlu is causing problems. In the future we will
need to be able to build against more tpls for the installation testing, but for
now we are only doing a very simple build.