Ifpack2,Tpetra: Removed unneeded zero fill of init vec in relaxations.
Ifpack2's Jacobi, Gauss-Seidel, and Symmetric Gauss-Seidel (all in the
Ifpack2::Relaxation class) give users the option to fill the initial
guess multivector (MV) with zeros.  Ifpack2 was doing the fill with
zeros redundantly in some cases, for all three relaxation methods.
Furthermore, telling the (S)GS implementation in Tpetra::CrsMatrix to
do the fill, rather than doing it in Ifpack2, passes down more info to
the implementation that it can use to avoid an initial Import.  (it is
not necessary to do an Import if the MV is full of zeros anyway.)
Since the common case of Gauss-Seidel is to do a small number of
sweeps with a zero initial guess, any redundant Import or zero fill
operations may affect performance noticeably.

This commit first modifies the (S)GS interface in Tpetra::CrsMatrix by
passing down to it the "zero initial guess" option.  Tpetra then uses
that option to decide whether or not to fill with zeros, as well as
whether or not to do an initial Import.  This is a backwards
incompatible change to the interface of Tpetra::CrsMatrix, but it is
justified in that it avoids the initial Import.  For a single (S)GS
sweep, this reduces the number of Import operations from one to zero,
giving Gauss-Seidel the same interprocess communication cost as one
sweep of Jacobi.  This method is also new in Trilinos 11.x, so we have
not imposed backwards compatibility restrictions on ourselves yet.
Furthermore, the only known client of this method is Ifpack2.  MueLu
only accesses Tpetra's Gauss-Seidel through Ifpack2 and should never
use Tpetra's Gauss-Seidel interface directly.

This commit also removes the initial zero fill from Ifpack2's Jacobi.
In fact, the only paths in Ifpack2::Relaxation that now do an initial
zero fill (in Ifpack2 itself, rather than implicitly or in Tpetra) are
the unoptimized RowMatrix versions of (Symmetric) Gauss-Seidel.

Note that this commit does not change the "in-place" version of (S)GS
in Tpetra::CrsMatrix (gaussSeidel), only the "copy" version of (S)GS
(gaussSeidelCopy).  Ifpack2 only calls the latter version.  It would
be straightforward to apply this optimization to the in-place version,
but since this would not affect Ifpack2 performance, I have deferred
this change for now.

Build/Test Cases Summary
Enabled Packages: Teuchos, Kokkos, Tpetra, Belos, Ifpack2
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=243,notpassed=0 (11.63 min)
3) SERIAL_RELEASE_SS => passed: passed=224,notpassed=0 (12.94 min)