Tpetra: DistObject: Renamed removeEmptyProcesses; added nonmember fns.
This commit renames removeEmptyProcesses to removeEmptyProcessesInPlace.
This should help avoid confusion with Map's removeEmptyProcesses method,
which (unlike removeEmptyProcessesInPlace) does _not_ modify in place.

This commit also adds removeEmptyProcessesInPlace nonmember functions
for DistObject.  Users are meant to call these nonmember functions
instead of the (not public) member functions of the same name of
DistObject.  The nonmember functions take an RCP<DistObject>& (nonconst
RCP reference), which gets invalided (set to null) on excluded
processes.  This prevents a "zombie" DistObject being left behind for
users to clean up, as long as the reference count of the RCP on input is
one.  (We could throw if it's not one, but that would require another
all-reduce over the original communicator, since different processes
might have different reference counts.)