make MINRES' setParameters() behave like the other kids
The way that Belos solver parmeters are handled at the moment is that the
solver instance is initially set up with a set of default parameters, and each
call to setParameters(new_params) sets what is given in new_params, everything
else remains the way it was. Note that this behavior makes Belos Krylov solvers
stateful.
The MINRES solver was the only one with a different approach: Upon
setParameters(new_params), all parameters in new_params would get set, and all
valid parameters not present in new_params would get set to their default
values (instead of just not touching them at all). This is stateless behavior.
While statelessness is certainly desirable in this situation, these two
different approaches caused trouble in all applications interfacing Belos'
MINRES, e.g., Stratimikos.
This commit adapts MINRES' setParameters() such that it behaves like the
other kids on the block.