MueLu: Xpetra: fixed a problem with a reference to a temporal object
MueLu Tpetra adapter implements the interface of the Tpetra Operator. As
such, the calls to getDomainMap() and getRangeMap() must return
references to RCP. That means, that we have to cast the corresponding
Xpetra::Map to the const reference. However, this was impossible due to
the fact that the corresponding cast function could also deal with zero
maps, which obviously cannot be referenced.
I solved the problem by adding an extra function to Xpetra which returns
the reference, but assumes that the map is not zero.