Tpetra: Map: Removed a reduction from noncontiguous Map constructor.
Tpetra::Map's "noncontiguous" constructor (that takes an
ArrayView<const GO> of GIDs owned by the calling process) was doing
an all-reduce to figure out whether the Map is distributed or locally
replicated.  "Distributed" means that there is at least one process
which owns strictly fewer elements than the global number of elements
in the Map.  It turns out that when GO (GlobalOrdinal) is a signed
integer type (the common case -- either int, long, or long long), this
all-reduce can be combined with the previous max all-reduce that
computes the global min and max GIDs.  This commit applies this
optimization.

I've also changes the Packet type for Boolean all-reduces in Map from
char to int.  Teuchos::reduceAll has a specialization for Packet=int
that should make those all-reduces faster.  (It skips over Teuchos::Comm
serialization and uses the predefined MPI_Op, rather than serializing
and using a custom MPI_Op.)  There is no specialization yet for Packet=
char because I couldn't get that specialization to compile correctly a
while back.