Piro: Enabled generic observers in SolverFactory
The two supported solvers both accept observers that implement the
Piro::ObserverBase interface, and the goal is to have every Thyra-based
Piro solver do so, hence the `generic' qualifier.
Therefore the generic observer instance can be directly passed as an
(optional) input argument to the createSolver function instead of
requiring a more involved mechanism.