Panzer: Enabling tpetra through model evaluator factory interface
A few things had to be added to get this working. In particular the
linear object factory needed a begin fill and end fill call for use
by the assembly engine. I also need to remove the epetra specifics
from the nox observer.

Also modified the solveTpetraSystem in the curl laplacian example.

Fixed several tests that were failing as a result of resumeFill/fillComplete
calls.