Tpetra: Added new test executable for Matrix Market I/O.
The test executable will help me ensure that Matrix Market I/O works
correctly for test matrices stored in a file, that are to be read in
using a row Map stored in a file.  The test is self contained; it
checks whether writing out the read-in Map and reading it back in
again results in the same Map, and ditto for the read-in CrsMatrix.
It is meant for real-valued matrices.  The test only builds by default;
it does not actually run as part of the check-in test suite, because I
want to test it on demand with large files that come from real
applications.

This commit also removes an old "test" for Matrix Market I/O that was
like this new test executable (it had to be run manually), but was less
effective (it didn't automatically test the written-out and read-in
matrix for equality with the original read-in matrix).

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => passed: passed=53,notpassed=0 (1.18 min)
1) SERIAL_RELEASE => passed: passed=41,notpassed=0 (0.02 min)