MueLu: CoalesceDropFactory: fixing the construction of the amalgamated graph
In the process of construction, we need to translate LIDs of the
original column map into LIDs of the amalgamated column map. The
algorithm we used had some implicit assumption built into it, such as
original column map GIDs are sorted and that GIDs in each block are all
present and consecutive.

This commit removes both of these assumption. Now we should be able to
work with any column map.