Tpetra: Matrix Market I/O: Got test to cooperate w/ unit test framework.
The unit test now fails when it is supposed to fail.  It apparently
has to set the 'success' variable (silently created by the unit test
macro) to false in order for the test to fail.  This came up because
my unit test was calling a separate function, in which the 'success'
variable was not defined.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=172,notpassed=0 (9.24 min)
3) SERIAL_RELEASE_SS => passed: passed=156,notpassed=0 (0.49 min)
4) MPI_RELEASE_SS => passed: passed=172,notpassed=0 (13.37 min)
Other local commits for this build/test group: 7b46c70, 8c90b9e, 935bec7, a29f2f4