Tpetra: Import: Added method to compute union of two Import objects.
If two Import object have the same source Map, then the _union_ of the
two objects is defined as the Import that would result from the union
of their (possibly different) target Maps.  The union of two Maps is
the Map whose GID list on each process is the union (in a possibly
different order) of the two Maps' GID lists on that process.

Computing the union of two Import objects is an important part of
adding two sparse matrices together, if both of them have column Maps
and Import objects.

This initial implementation is not optimized, but it works.  This
commit also includes a unit test for the union operation.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=125,notpassed=0 (3.21 min)
3) SERIAL_RELEASE_SS => passed: passed=110,notpassed=0 (0.31 min)
4) MPI_RELEASE_SS => passed: passed=125,notpassed=0 (2.15 min)