Disabling PyTrilinos if not enabled and BUILD_SHARED_LIBS=OFF
This logic is similar to logic that exists for ForTrilinos and TriKota.  If
Fortran is not enabled, then ForTrilinos is turned off.  In that case, it is
hard turned off, even if ForTrilinos was explicitly enabled.  That is likely
bad that if you explicitly turn on ForTrilinos but don't have Fortran then it
will implicity disable ForTrilinos but that is how we did this.

In the case of PyTrilinos, we only disable it if it is not already enabled and
BUILD_SHARED_LIBS is off.  If PyTrilinos is explicitly enabled then PyTrilinos
will fail to configure, just as it did before so we maintain backward
compatibility.  I can't see how a user would be upset

In general we need to make it so that Trilinos configures in a more inteligent
way for large sets of package, especially in automated testing.  We need to
put in this type of logic for more packages likely.  Note that if "shared
libs" was treated as a TPL in the TriBITS system (not saying that we should),
then PyTrilinos would be auotmatically disabled unless "shared libs" was
enabled.  A change like this just makes that behavior a little more
consistent.