Tpetra: Map: Exclude initial sequence of contiguous GIDs from table.
Map already optimizes GID -> LID lookups by first checking if a GID is
in the initial contiguous sequence of GIDs, and only going to the GID
-> LID lookup table if it is not.  Nevertheless, Map was still storing
that initial contiguous sequence of GIDs in the hash table, even
though it was never looking up those GIDs.  This was wasting Map
construction time and storage space.  Depending on the hash function
and the number and arrangement of noncontiguous GIDs, it also might
slow down GID -> LID lookups, by reducing locality.  (Keeping the
table small makes cache misses less likely.)

This commit changes Map construction so that only the GIDs following
the initial noncontiguous sequence get stored in the GID -> LID table.
It also includes an improved describe() method for FixedHashTable (the
"fixed-structure" hash table that Map now uses) that dumps the table's
contents in YAML format.