MueLu: Fixing bug in ArbitrateAndCommunicate.
One of this method's purposes is to provide tie-breaking among processors that doesn't depend on the PID.
It uses pseudo-random weights that are not supposed to be correlated across processors.  However, A&C was
using Tpetra::Vector::random(), which produces exactly the same random values locally across processors.
(To be fair, this is well-documented.)  So it was defaulting to using the highest PID for tie-breaking.

The result was an excessively high number of iterations in DeterminePartitionPlacement, which attempts to
assign partitions from rebalancing to reduce data communication.

On a milestone test problem on 256 cores, here's the effect:

Before patch:

    Level 1
    DeterminePartitionPlacement: Arbitration phase
    Number of arbitration rounds = 18
    Level 2
    DeterminePartitionPlacement: Arbitration phase
    Number of arbitration rounds = 9

After patch:

    Level 1
    DeterminePartitionPlacement: Arbitration phase
    Number of arbitration rounds = 6

    Level 2
    DeterminePartitionPlacement: Arbitration phase
    Number of arbitration rounds = 4