MueLu: disable ML multiply when using experimental code
Energy minimization uses AP pattern for restriction. The problem with ML multiply is that it automatically
removes zero valued entries in the matrix product, resulting in incorrect pattern for the minimization.
One could try to mitigate that by multiply matrices with all entries equal to zero, which would produce the
correct graph. However, that is one extra MxM we don't need.

The current solution is to disable ML multiply when experimental option is specified.