Tpetra: CrsMatrix: added override of default impl of RowMatrix::add.
CrsMatrix now overrides the default implementation of RowMatrix:add.
For now, the override only differs in that it ensures the returned
RowMatrix is actually a CrsMatrix of the same type as *this (including
the fifth template parameter).  The default implementation in
RowMatrix does return a CrsMatrix, but it assumes the default fifth
template parameter of CrsMatrix.  Returning a CrsMatrix of the same
type as *this avoids potential downstream issues with dynamic casts
failing.  We plan to optimize CrsMatrix::add more later.