Kokkos : UQ/Ensemble hybrid parallel CG-solve miniapp, fix deep_copy for subviews, allow 'operator()' to have extra unused indices.
Allow View::operator() to have extra ignored integer arguments so that
computations can be generalized to the maximum rank with minimum
performance penalty.

Initial testing of hybrid parallel CG solve of UQ/ensemble linear system
where the sparse matrix graph is shared and the matrix and vector
values are ensembles.