Tribits: Fix issue with export system
The recent refactor of the export system changed a variable name that was used
to define where the include directories were. The new name was more consistent
with the other names used to create the configure files so the template files
were updated to use it instead of resurecting the old name.