ML: Adding support for ParMETIS 4.x & METIS 5.x.
This commit allows ML to be used with ParMETIS 4.x and METIS 5.x -- the latter comes with the
former.  There have been multiple requests for this, the latest from Xyce. Actually, Xyce just
wanted ML to compile in the presence of the newer versions, but since it wasn't much more effort,
I just made it work.  The main issue was that METIS (ParMETIS) broke backwards compatibility from
4.x to 5.x (3.x to 4.x).

I've tried bunch of different compile configurations.

    no METIS, no ParMETIS
    ParMETIS 4.x, METIS 5.x
    ParMETIS 4.x, no METIS
    no ParMETIS, METIS 5.x
    no ParMETIS, METIS 4.x
    ParMETIS 3.x, no METIS
    no ParMETIS, METIS 4.x

There are no additional tests, but the existing ML METIS test is still passing.