MueLu: Emin: removed rhs vector (nullspace) from iterative solvers
We use iterative solvers to compute an update to the initial guess.
Therefore, this update is in the affine space, where it satisfies
nonzero pattern and XdP = 0, instead of XdP = B. Therefore, there is no
need to require B in the iterative methods, as we don't use it. In fact,
it is contained implicitly in the constraint application.